functions to manage qgis projects in QgsPostgresUtils
