+++ src/providers/postgres/qgspostgresutils.h
@@ -97,6 +97,33 @@ class QgsPostgresUtils
     static bool columnExists( QgsPostgresConn *conn, const QString &table, const QString &column );
 
     static bool tableExists( QgsPostgresConn *conn, const QString &name );
+
+    /*
+    * Check if projects table exists in the specified schema
+    *
+    * \returns true if the table exists
+    * 
+    * \since QGIS 3.44
+    */
+    static bool projectsTableExists( QgsPostgresConn *conn, const QString &schemaName );
+
+    /*
+    * Creates projects table exists in the specified schema
+    *
+    * \returns true on success
+    * 
+    * \since QGIS 3.44
+    */
+    static bool createProjectsTable( QgsPostgresConn *conn, const QString &schemaName );
+
+    /*
+    * Deletes projects from project's table in the specified schema
+    *
+    * \returns true on success
+    * 
+    * \since QGIS 3.44
+    */
+    static bool deleteProjectFromSchema( QgsPostgresConn *conn, const QString &projectName, const QString &schemaName );
 };
 
 #endif
